refactor: consolidate UnionScopes and ScopesFromToken into shared auth package
